+++ extensions/lorem-ipsum/src/utils.tsx
@@ -23,11 +23,28 @@ export const generateParagraphs = (count: number) => {
       generator.generateSentences(
         Math.floor(
           Math.random() *
-            (loremIpsumOptions.sentencesPerParagraph.max - loremIpsumOptions.sentencesPerParagraph.min + 1),
-        ) + loremIpsumOptions.sentencesPerParagraph.min,
-      ),
+            (loremIpsumOptions.sentencesPerParagraph.max - loremIpsumOptions.sentencesPerParagraph.min + 1)
+        ) + loremIpsumOptions.sentencesPerParagraph.min
+      )
     )
     .join("\r\n\r\n"); // newline + seperator line
+};
+
+export const generateHtmlParagraphs = (count: number) => {
+  return (
+    "<p>" +
+    Array.from(Array(count))
+      .map(() =>
+        generator.generateSentences(
+          Math.floor(
+            Math.random() *
+              (loremIpsumOptions.sentencesPerParagraph.max - loremIpsumOptions.sentencesPerParagraph.min + 1)
+          ) + loremIpsumOptions.sentencesPerParagraph.min
+        )
+      )
+      .join("</p>\r\n\r\n<p>") +
+    "</p>"
+  ); // newline + seperator line
 };
 
 export const generateSentences = (count: number) => {